+++ src/server/cray/cfs/api/controllers/sessions.py
@@ -1,7 +1,7 @@
 #
 # MIT License
 #
-# (C) Copyright 2019-2024 Hewlett Packard Enterprise Development LP
+# (C) Copyright 2019-2026 Hewlett Packard Enterprise Development LP
 #
 # Permission is hereby granted, free of charge, to any person obtaining a
 # copy of this software and associated documentation files (the "Software"),
@@ -48,10 +48,23 @@ CONFIG_DB = dbutils.get_wrapper(db='configurations')
 _kafka = None
 
 
+class JobFieldAlreadySet(Exception):
+    """
+    CASMCMS-9627: Raised when attempting to patch a status.session.job field when it is already set
+    """
+    def __init__(self, patch_job: str, actual_job: str) -> None:
+        super().__init__(
+            "status.session.job field cannot be updated after it has been set; "
+            f"current value: {actual_job}, patch value: {patch_job}"
+        )
+
+
 def _init(topic='cfs-session-events'):
     """ Initialize the kafka producer information """
     global _kafka
+    LOGGER.debug("_init: Initializing ProducerWrapper")
     _kafka = kafka_utils.ProducerWrapper(topic)
+    LOGGER.debug("_init: ProducerWrapper initialized")
 
 
 @dbutils.redis_error_handler
@@ -118,7 +131,10 @@ def create_session_v2():  # noqa: E501
     # end workaround
     session_data['status']['session']['start_time'] = datetime.datetime.now().isoformat(timespec='seconds')
     _kafka.produce(event_type='CREATE', data=session_data)
-    response_data = DB.put(session_data['name'], session_data)
+    session_name = session_data['name']
+    LOGGER.debug("create_session_v2: Writing new session '%s' to database", session_name)
+    response_data = DB.put(session_name, session_data)
+    LOGGER.debug("create_session_v2: DB put complete for '%s'", session_name)
     return convert_session_to_v2(response_data), 200
 
 
@@ -177,7 +193,10 @@ def create_session_v3():  # noqa: E501
     data = session.to_dict()
     data['status']['session']['start_time'] = datetime.datetime.now().isoformat(timespec='seconds')
     _kafka.produce(event_type='CREATE', data=data)
-    response_data = DB.put(data['name'], data)
+    session_name = data['name']
+    LOGGER.debug("create_session_v3: Writing new session '%s' to database", session_name)
+    response_data = DB.put(session_name, data)
+    LOGGER.debug("create_session_v3: DB put complete for '%s'", session_name)
     _set_link(response_data)
     return response_data, 201
 
@@ -234,7 +253,9 @@ def delete_session_v2(session_name):  # noqa: E501
             status=404, title="Session not found.",
             detail="Session {} could not be found".format(session_name))
     session = DB.get(session_name)
+    LOGGER.debug("delete_session_v2: Deleting '%s' in database", session_name)
     DB.delete(session_name)
+    LOGGER.debug("delete_session_v2: Deleted '%s' in database", session_name)
     _kafka.produce(event_type='DELETE', data=session)
     return None, 204
 
@@ -257,7 +278,9 @@ def delete_session_v3(session_name):  # noqa: E501
             status=404, title="Session not found.",
             detail="Session {} could not be found".format(session_name))
     session = DB.get(session_name)
+    LOGGER.debug("delete_session_v3: Deleting '%s' in database", session_name)
     DB.delete(session_name)
+    LOGGER.debug("delete_session_v3: Deleted '%s' in database", session_name)
     _kafka.produce(event_type='DELETE', data=session)
     return None, 204
 
@@ -304,7 +327,9 @@ def delete_sessions_v2(age=None,  min_age=None, max_age=None,
                                                   succeeded=succeeded, tag_list=tag_list)
         for session in sessions_data:
             session_name = session['name']
+            LOGGER.debug("delete_sessions_v2: Deleting '%s' in database", session_name)
             DB.delete(session_name)
+            LOGGER.debug("delete_sessions_v2: Deleted '%s' in database", session_name)
             _kafka.produce(event_type='DELETE', data=session)
     except ParsingException as err:
         return connexion.problem(
@@ -480,8 +505,7 @@ def get_sessions_v3(age=None, min_age=None, max_age=None, status=None, name_cont
 def patch_session_v2(session_name):
     """Update a Config Framework Session
 
-    Updates a new V2Session # noqa: E501
-
+    Updates a V2Session # noqa: E501
     :rtype: V2Session
     """
     LOGGER.debug("PATCH /v2/sessions/id invoked patch_session")
@@ -500,7 +524,11 @@ def patch_session_v2(session_name):
             status=404, title="Session not found.",
             detail="Session {} could not be found".format(session_name))
     data = dbutils.convert_data_from_v2(data, V2Session)
-    response_data = _patch_session(session_name, data)
+    # CASMCMS-9627: To minimize changes, only update the V3 API.
+    # This is fine because that is what cfs-operator uses. This also allows
+    # a way for an admin to bypass the restrictions, if for some reason it is ever
+    # needed.
+    response_data = _patch_session(session_name, data, job_update_restrictions=False)
     return convert_session_to_v2(response_data), 200
 
 
@@ -509,7 +537,9 @@ def patch_session_v2(session_name):
 def patch_session_v3(session_name):
     """Update a Config Framework Session
 
-    Updates a new V3Session # noqa: E501
+    Updates a V3Session # noqa: E501
+
+    If the job field is being patched, return a 409 if the job field has already been set
 
     :rtype: V3Session
     """
@@ -528,7 +558,13 @@ def patch_session_v3(session_name):
         return connexion.problem(
             status=404, title="Session not found.",
             detail="Session {} could not be found".format(session_name))
-    response_data = _patch_session(session_name, data)
+    try:
+        response_data = _patch_session(session_name, data, job_update_restrictions=True)
+    except JobFieldAlreadySet as err:
+        LOGGER.debug(err)
+        return connexion.problem(
+            status=409, title="Session patch conflict.",
+            detail=f"Session {session_name} could not be patched: {err}")
     return response_data, 200
 
 
@@ -540,11 +576,20 @@ STATUS_ORDERING = {
 }
 
 
-def _patch_session(session_name, new_data):
+def _patch_session(session_name, new_data, job_update_restrictions: bool):
+    """
+    Applies the patch_data to the specified session, and returns the updated session data.
+    If job_update_restrictions is true, call _enforce_job_update_restrictions.
+    """
     data = DB.get(session_name)
     status = data['status']
     artifacts = status['artifacts']
     session = status['session']
+
+    patch_session = new_data.get('status', {}).get('session', {})
+
+    if job_update_restrictions:
+        _enforce_job_update_restrictions(session, patch_session)
 
     # Artifacts
     for artifact in new_data.get('status', {}).get('artifacts', []):
@@ -558,7 +603,7 @@ def _patch_session(session_name, new_data):
             artifacts.append(artifact)  # No artifacts matched
 
     # Session Status
-    for key, value in new_data.get('status', {}).get('session', {}).items():
+    for key, value in patch_session.items():
         if value:  # Never overwrite with an empty field
             if key in STATUS_ORDERING:
                 ordering = STATUS_ORDERING[key]
@@ -572,6 +617,28 @@ def _patch_session(session_name, new_data):
                 session[key] = value
     return DB.put(session_name, data)
 
+
+def _enforce_job_update_restrictions(session_status_session_data,
+                                     patch_status_session_data) -> None:
+    """
+    Raise JobFieldAlreadySet exception if the patch is setting the job field after
+    it already has been set.
+
+    session_status_session_data = .status.session for the existing CFS session
+    patch_status_session_data = .status.session for the patch
+    """
+    try:
+        session_job = session_status_session_data["job"]
+        patch_job = patch_status_session_data["job"]
+    except KeyError:
+        # No problem if the field is not present in the existing CFS entry,
+        # or not in the patch data
+        return
+
+    # The field is present in both the CFS DB and the patch data. Raise an exception if
+    # the CFS DB field has already been set and the patch is trying to set it.
+    if session_job and patch_job:
+        raise JobFieldAlreadySet(patch_job, session_job)
 
 def _validate_session_target(target):
     """Validate the target section